requirements/base.txt
Django
psycopg2
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-crispy-forms
flake8
black